Fix build errors: DISALLOW_CONFIG_ACCESSIBILITY API 36+ compat, WebRTC type mismatch, disable lint abort

+++ app/src/main/kotlin/com/familyguardpro/webrtc/WebRTCClient.kt
@@ -953,7 +953,7 @@ class WebRTCClient(
                 }
             }
             lastBytesSent = bytesSent
-            lastStatsTimestamp = timestamp
+            lastStatsTimestamp = timestamp.toLong()
             
             // Calculate packet loss percentage
             val totalPackets = packetsLost + packetsSent
